backend/requirements.txt
# ═══════════════════════════════════════════════════════════════════════
# Face Attendance System — Python 3.11 Requirements
# ═══════════════════════════════════════════════════════════════════════

# ── Web Framework ────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────
sqlmodel==0.0.18
# pydantic v2 is pulled in by sqlmodel

# ── Settings ──────────────────────────────────────────────────────────────
pydantic-settings==2.2.1

# ── Auth / JWT ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0

# ── CV: Core ──────────────────────────────────────────────────────────────
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0

# ── CV: Face Detection & Recognition (ArcFace) ───────────────────────────
# InsightFace provides RetinaFace detection + ArcFace recognition in one package
insightface==0.7.3
onnxruntime==1.17.3          # CPU inference backend for InsightFace models

# ── CV: Landmark-based Liveness (EAR blink, nod, smile) ─────────────────
mediapipe==0.10.13

# ── CV: Fast Embedding Search ─────────────────────────────────────────────
faiss-cpu==1.8.0             # Facebook AI Similarity Search (CPU build)

# ── Utilities ─────────────────────────────────────────────────────────────
httpx==0.27.0                # For testing endpoints
python-dotenv==1.0.1
